fix misskey-dev/misskey-#10417

+++ packages/backend/src/queue/processors/ImportCustomEmojisProcessorService.ts
@@ -86,6 +86,10 @@ export class ImportCustomEmojisProcessorService {
 					continue;
 				}
 				const emojiInfo = record.emoji;
+				if (!/^[a-zA-Z0-9_]+$/.test(emojiInfo.name)) {
+					this.logger.error(`invalid emojiname: ${emojiInfo.name}`);
+					continue;
+				}
 				const emojiPath = outputPath + '/' + record.fileName;
 				await this.emojisRepository.delete({
 					name: emojiInfo.name,
